+++ src/routes/finance/+page.ts
@@ -2,5 +2,5 @@ import { redirect } from '@sveltejs/kit';
 import { resolve } from '$app/paths';
 
 export function load() {
-	redirect(307, resolve('/finance/bills'));
+	redirect(307, resolve('/finance/income'));
 }